git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@951 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/ASPHERE/atom_vec_ellipsoid.cpp
@@ -583,9 +583,9 @@ int AtomVecEllipsoid::data_vel_hybrid(int m, char **values)
    return # of bytes of allocated memory
 ------------------------------------------------------------------------- */
 
-int AtomVecEllipsoid::memory_usage()
+double AtomVecEllipsoid::memory_usage()
 {
-  int bytes = 0;
+  double bytes = 0.0;
 
   if (atom->memcheck("tag")) bytes += nmax * sizeof(int);
   if (atom->memcheck("type")) bytes += nmax * sizeof(int);
