+++ src/DIPOLE/atom_vec_dipole.cpp
@@ -638,9 +638,9 @@ int AtomVecDipole::data_vel_hybrid(int m, char **values)
    return # of bytes of allocated memory 
 ------------------------------------------------------------------------- */
 
-int AtomVecDipole::memory_usage()
+double AtomVecDipole::memory_usage()
 {
-  int bytes = 0;
+  double bytes = 0.0;
 
   if (atom->memcheck("tag")) bytes += nmax * sizeof(int);
   if (atom->memcheck("type")) bytes += nmax * sizeof(int);
