+++ src/GRANULAR/fix_wall_gran.h
@@ -27,7 +27,7 @@ class FixWallGran : public Fix {
   void setup();
   void post_force(int);
 
-  int memory_usage();
+  double memory_usage();
   void grow_arrays(int);
   void copy_arrays(int, int);
   int pack_exchange(int, double *);
