+++ src/MOLECULE/angle_hybrid.cpp
@@ -251,9 +251,9 @@ double AngleHybrid::single(int type, int i1, int i2, int i3)
    memory usage
 ------------------------------------------------------------------------- */
 
-int AngleHybrid::memory_usage()
+double AngleHybrid::memory_usage()
 {
-  int bytes = 0;
+  double bytes = 0.0;
   for (int m = 0; m < nstyles; m++) bytes += maxangle[m]*4 * sizeof(int);
   for (int m = 0; m < nstyles; m++) 
     if (styles[m]) bytes += styles[m]->memory_usage();
