+++ src/MOLECULE/atom_vec_angle.cpp
@@ -631,9 +631,9 @@ int AtomVecAngle::data_atom_hybrid(int nlocal, char **values)
    return # of bytes of allocated memory 
 ------------------------------------------------------------------------- */
 
-int AtomVecAngle::memory_usage()
+double AtomVecAngle::memory_usage()
 {
-  int bytes = 0;
+  double bytes = 0.0;
 
   if (atom->memcheck("tag")) bytes += nmax * sizeof(int);
   if (atom->memcheck("type")) bytes += nmax * sizeof(int);
