+++ src/MOLECULE/atom_vec_bond.cpp
@@ -572,9 +572,9 @@ int AtomVecBond::data_atom_hybrid(int nlocal, char **values)
    return # of bytes of allocated memory 
 ------------------------------------------------------------------------- */
 
-int AtomVecBond::memory_usage()
+double AtomVecBond::memory_usage()
 {
-  int bytes = 0;
+  double bytes = 0.0;
 
   if (atom->memcheck("tag")) bytes += nmax * sizeof(int);
   if (atom->memcheck("type")) bytes += nmax * sizeof(int);
