+++ src/MOLECULE/atom_vec_molecular.cpp
@@ -766,9 +766,9 @@ int AtomVecMolecular::data_atom_hybrid(int nlocal, char **values)
    return # of bytes of allocated memory 
 ------------------------------------------------------------------------- */
 
-int AtomVecMolecular::memory_usage()
+double AtomVecMolecular::memory_usage()
 {
-  int bytes = 0;
+  double bytes = 0.0;
 
   if (atom->memcheck("tag")) bytes += nmax * sizeof(int);
   if (atom->memcheck("type")) bytes += nmax * sizeof(int);
