+++ src/MOLECULE/dihedral_hybrid.cpp
@@ -248,9 +248,9 @@ void DihedralHybrid::read_restart(FILE *fp)
    memory usage
 ------------------------------------------------------------------------- */
 
-int DihedralHybrid::memory_usage()
+double DihedralHybrid::memory_usage()
 {
-  int bytes = 0;
+  double bytes = 0.0;
   for (int m = 0; m < nstyles; m++) bytes += maxdihedral[m]*5 * sizeof(int);
   for (int m = 0; m < nstyles; m++) 
     if (styles[m]) bytes += styles[m]->memory_usage();
