+++ src/MOLECULE/improper_hybrid.cpp
@@ -235,9 +235,9 @@ void ImproperHybrid::read_restart(FILE *fp)
    memory usage
 ------------------------------------------------------------------------- */
 
-int ImproperHybrid::memory_usage()
+double ImproperHybrid::memory_usage()
 {
-  int bytes = 0;
+  double bytes = 0.0;
   for (int m = 0; m < nstyles; m++) bytes += maximproper[m]*5 * sizeof(int);
   for (int m = 0; m < nstyles; m++) 
     if (styles[m]) bytes += styles[m]->memory_usage();
