+++ src/atom_vec_atomic.cpp
@@ -444,9 +444,9 @@ int AtomVecAtomic::data_atom_hybrid(int nlocal, char **values)
    return # of bytes of allocated memory 
 ------------------------------------------------------------------------- */
 
-int AtomVecAtomic::memory_usage()
+double AtomVecAtomic::memory_usage()
 {
-  int bytes = 0;
+  double bytes = 0.0;
 
   if (atom->memcheck("tag")) bytes += nmax * sizeof(int);
   if (atom->memcheck("type")) bytes += nmax * sizeof(int);
