+++ src/compute_epair_atom.cpp
@@ -224,8 +224,8 @@ void ComputeEpairAtom::unpack_reverse_comm(int n, int *list, double *buf)
    memory usage of local atom-based array
 ------------------------------------------------------------------------- */
 
-int ComputeEpairAtom::memory_usage()
+double ComputeEpairAtom::memory_usage()
 {
-  int bytes = nmax * sizeof(double);
+  double bytes = nmax * sizeof(double);
   return bytes;
 }
