+++ src/compute_variable_atom.cpp
@@ -87,8 +87,8 @@ void ComputeVariableAtom::compute_peratom()
    memory usage of local atom-based array
 ------------------------------------------------------------------------- */
 
-int ComputeVariableAtom::memory_usage()
+double ComputeVariableAtom::memory_usage()
 {
-  int bytes = nmax * sizeof(double);
+  double bytes = nmax * sizeof(double);
   return bytes;
 }
