+++ src/fix.h
@@ -100,7 +100,7 @@ class Fix : protected Pointers {
 
   virtual int modify_param(int, char **) {return 0;}
 
-  virtual int memory_usage() {return 0;}
+  virtual double memory_usage() {return 0.0;}
 };
 
 }
