+++ src/kspace.h
@@ -37,7 +37,7 @@ class KSpace : protected Pointers {
   virtual void setup() = 0;
   virtual void compute(int, int) = 0;
   virtual void timing(int, double &, double &) {}
-  virtual int memory_usage() {return 0;}
+  virtual double memory_usage() {return 0.0;}
 };
 
 }
